requirements.txt
alembic==1.15.1
aniso8601==10.0.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
